+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-# Unreleased
+## Unreleased
 
+- Added test for `retried` field on transaction.sale response
 - Add ExchangeRateQuote API
 
 ## 3.10.0